sample-code/image-processing/searchable-pdf/requirements.txt
fdk
oci
wheel
pillow
pytesseract
opencv-python